requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
openai>=1.0.0
chromadb>=0.4.0
pypdf>=3.0.0
openpyxl>=3.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.0
cohere>=5.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
